requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
qrcode[pil]==7.4.2